requirements.txt
# Web framework
fastapi
uvicorn

# Deep learning & Transformers
torch
transformers

# Image handling
pillow

# CORS middleware
python-multipart

# Optional: only if you want to use OpenCV for debugging or webcam demo
opencv-python

flask
